experience_bonus_mult bug fix
Error: Json(Error("invalid type: floating point `1.15`, expected u64", line: 1, column: 191630))
Causes a JSON error because it can be a floating point but is declared as an integer
This is backed up because of the value of ExperienceBonusMult in my profile's JSON is 1.15.

+++ src/profile.rs
@@ -310,7 +310,7 @@ pub struct Stats {
     /// Session experience multiplier?
     pub session_experience_mult: f64,
     /// Experience bonus multiplier?
-    pub experience_bonus_mult: u64,
+    pub experience_bonus_mult: f64,
     /// Total session experience
     pub total_session_experience: u64,
     /// Last session timestamp
